Fix fontconfig non-family defaults
Main problem was that FcPatternRemove was used instead of
FcPatternDel.  Remove only removes one value whereas we wanted
to remove all family values.

After that, if we where to add a dummy value for FC_FAMILY to
query_without_family, then the code would work.  The reason we
need to add this is that if the query_without_family does NOT
have any FC_FAMILY, then FcFontRenderPrepare will copy the first
value of FC_FAMILY from query to query_without_family.  We don't
want that.

Alternatively, we can call FcConfigSubstituteWithPat instead
of FcFontRenderPrepare as this does exactly what we want: run
configuration stanzas with target=pattern.